Changed type of Id to Long in GreenBox and dependents.

+++ src/main/java/com/greenbox/test/demo/repository/GreenBoxRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GreenBoxRepository extends JpaRepository<GreenBox, Integer> {
+public interface GreenBoxRepository extends JpaRepository<GreenBox, Long> {
     List<GreenBox> findAllByUserId(Long userId); // Integer или Long?
 }
