+++ src/main/java/com/greenbox/test/demo/service/GreenBoxService.java
@@ -16,17 +16,17 @@ public interface GreenBoxService {
     /**Возвращает greenBox по его ID
      * @param id - ID greenBox
      * @return - объект greenBox с заданным ID*/
-    GreenBox read(int id);
+    GreenBox read(Long id);
 
     /**Обновляет greenBox с заданным ID,
      * в соответствии с переданным greenBoxом
      * @param greenBox - greenBox в соответсвии с которым нужно обновить данные
      * @param id - id greenBox которого нужно обновить
      * @return - true если данные были обновлены, иначе false*/
-    boolean update(GreenBox greenBox, int id);
+    boolean update(GreenBox greenBox, Long id);
 
     /**Удаляет greenBox с заданным ID
      * @param id - id greenBox, которого нужно удалить
      * @return - true если greenBox был удален, иначе false*/
-    boolean delete(int id);
+    boolean delete(Long id);
 }
